Move app elements with changes to keep them functional.
