+++ core/src/main/java/com/hfad/palamarchuksuperapp/core/ui/composables/basic/AppTextField.kt
@@ -18,8 +18,8 @@ import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.hfad.palamarchuksuperapp.ui.compose.theme.AppTheme
-import com.hfad.palamarchuksuperapp.R
+import com.hfad.palamarchuksuperapp.core.R
+import com.hfad.palamarchuksuperapp.core.ui.theme.AppTheme
 
 /**
  * Компонент для редактирования текста с поддержкой метки и подсказки.
@@ -109,7 +109,7 @@ internal fun AppTextFieldPreview() {
                     onValueChange = { text.value = it },
                     label = {
                         AppText(
-                            R.string.model_hint,
+                            R.string. place_holder_text,
                         )
                     },
                 )
@@ -126,7 +126,7 @@ internal fun AppTextFieldPreview() {
                     onValueChange = { text.value = it },
                     label = {
                         AppText(
-                            R.string.model_hint,
+                            R.string.place_holder_text,
                         )
                     },
                 )
